+++ KTN House/src/components/PostCard.css
@@ -7,5 +7,6 @@
   border: 1px solid #accfff;
   background-color: #accfff;
   cursor: pointer;
+  padding: 10px;
  
 }
