+++ KTN House/src/pages/Logins/Login.css
@@ -4,7 +4,7 @@
   justify-content: center;
   padding: 80px 0;
 }
-.col-login {
+.colo-login {
   background-color: rgb(252, 250, 250);
   width: 400px;
   padding: 10px 24px;
@@ -17,7 +17,7 @@
   margin-bottom: 10px;
   padding-left: 10px;
 }
-.col-login h3 {
+.colo-login h3 {
   font-size: 20px;
   color: black;
   margin-bottom: 16px;
