+++ KTN House/src/pages/about.css
@@ -65,7 +65,7 @@
 .abt-time {
   background-color: rgb(226, 223, 223);
   width: 600px;
-  height: 150px;
+  height: 200px;
   padding: 10px 24px;
   border-radius: 4px;
   justify-content: space-around;
