+++ KTN House/src/topbar/Topbar.css
@@ -63,7 +63,11 @@ nav {
   color: #dbdcdc;
   cursor: pointer;
 }
-
+.logo {
+  font-weight: 900;
+  color: white;
+  margin-top: 15px;
+}
 .ktn {
   font-weight: 800;
 }
